1248 Border Regions 4
The final installment for adding borders to Charted Space in 1248. The only comments note in this package is that Ley Sector and Langere have been named Ley and Langer per the 1248 map with appropriate annotation additions in the xml files. This is in keeping with standard maintained to this point of keeping things as used in the source material.