1120 sectors 2 (#115)
Adding Diaspora Sector and Foreven, incorporating T509 data and new comments. Diaspora file name is correct as it represents 1124 allegiances but as source states is representative of the entire Rebellion until Hard Times. Also tweaked main xml. There is a mistake in Massilia. It was placed in position 1,0 when it should be in 0,1