Added Bellatrix, a star in Orion, to an otherwise barren area in Hadji
This is the approximate location of the star relative to Terra and the Orion OB1 Association, making it one of the few major Orion stars to lie in an already-charted sector. I'm not 100% sure where the stellar data should go in this data format, though, so I may just re-format the entire file to a newer format if this doesn't work properly.