Map: Reduce x/y animation duration to half of overall animation duration
With the previous animateTo algorithm, the target would often not end
up in view until towards the end of the animation, resulting in
extreme slew and lack of context, plus probably more tile reloads than
ideal.

Tweak the animation logic to only animate position for the first half,
so that the target becomes centered earlier. There's probably a
better/smoother algorithm, but this seems to improve things.