Map: Make Enter key do something useful in route box
With the route box showing, the Enter key would trigger the default
button, which was close. This could be confusing if the user had
specified a destination then hit Enter instead of clicking a J-n
button. Make it a no-op by adding a disabled default button.

And then to improve usability:

* Make Enter key in the route start field set focus to the route end
  field.

* Make Enter key in the route end field re-do the route search with
  the selected jump rating, or J-2 if none was selected.

Also, fix the width of the route box input fields, which seems to have
been accidentally changed by a find/replace in d4bb6f0a.