T5SS: Make res/t5ss/data the canonical source for T5SS data
Previous the master data was in a spreadsheet that included lots of
other fields, but was difficult to update.

Now the res/t5ss/data has the canonical data (which will sync with an
external store as well). The update_world_data.pl script is updated to
process the files in data, combine with the anomalies listing, and
update the M1105 data. It's round-about but it works. Hopefully.