Data Validation: Update RU calculation to match T5.10
In T5.10 the RU calculation is again given as R*L*I*E, but with the
caveat that any 0s are treated as 1s. The T5SS data did not have that
clause, so the lint validated the data. The data is being updated,
so the calc can now match the rules.