Path Finding: More robust A* algorithm
The previous A* algorithm was fine, but the weighting function would
yield routes that were asymmetric, e.g. A->B was more jumps than B->A.

This rewrites the algorithm following the code in Wikipedia (yay)
which makes tuning it easier. Right now the routes should be
symmetric, and minimal in terms of *jumps* but not necessarily
*parsecs*, which is important for minimizing fuel cost.

Tweaking that will be in a follow-up.

Also:
* Ensure TLS1.2 is available to be used
* Fix a null dereference when searching