Path Finding: Modify edge weights by jump distance.
A jump is a jump (so, 2 jumps is worse than 1), but there's a per-parsec
fuel cost, so bias the route by a small fraction (parsecs/36) to minimize
the fuel cost.

There's the possibility that over a long route that this would actually
add an extra jump to the route, but this is still better than the old
algorithm!

Resolves #253