Page: Send precise location requesting routes
This fixes an issue where sending two world names, e.g. "Inthe"
and "Aramis", is ambiguous for a location like the Spinward Marches.
When clicking or otherwise populating the route start/end fields,
the precise (sector/hex) location is stashed, and used rather
just the world name. If the fields are edited manually then the
typed name(s) are used.