MSEC API: Fall back to border/region code if no defined Allegiance
When serializing metadata to MSEC, the logic for outputting a color for
each border/region depended on the allegiance as input into a stylesheet
resolution call. If the allegiance was not defined, nothing would be
returned. This differs from the rendering logic which would fall back
to just using the code - so match that.