M1105: Move Holowon to "far away", restore Rim Worlds name
Years ago Brian Smaller submitted a sector from his ATU named "Rim
Worlds" that fit into the map where Holowon goes, a bit rimward of the
Solomani Rim. After some consideration, we're moving the sector out of
Charted Space and treating it like an ATU. The data's still there with
the new name, but Holowon is now undefined as far as the OTU goes.