M1105: Relegate Muarne & Lancask sectors to Turokan's Rim Expedition
A backburner project is to host all of the Rim Expedition sectors that
Bari Z. Stafford Sr (General Turokan) as an ATU as it doesn't really
fit the OTU for various reasons. Initially I'd included some of the
nearby sectors in Charted Space, but they're not well developed and
not a good fit. So removing them. The data is still on github if
anyone needs it, and some day it'll be possible to view the sectors as
an ATU.