Poster API: Add experimental "compositing" option
This is a simpler approach to what's in PR #337 to allow a sector
image to be clipped to the hex edge borders, which simplifies
overlaying a custom sector atop a full map image. This isn't in
a finished state, but will be iterated on before finalizing the
behavior and documenting it.