World Info: Show map source
A JSON file generated from the "Traveller World Details" is now used
to provide the source for world maps, shown below the thumbnails in
the world info panels.

The sheet and the map repo may be out of sync, so there may be maps
that don't get a source; we'll need to tackle that.