M1105: Fix size code for Kruse 3023 (Franklin)
The Mongoose text specifies size "S" for Franklin, but that is
used in some Traveller variants for "small world". The text reads:

"The mainworld of the otherwise strangely empty 0603 system is a
super-earth whose eccentric orbit centres on the barycentre of a pair
of binary gas giants. There are virtually no other bodies in the
system."

So made the size code F instead, which is the biggest size supported
in T5SS encoding - 6 times the mass of Earth, which is probably the
maximum size before you start getting a mini-Neptune instead.

Thanks to @CyberiaResurrection for pointing out the issue.