T5SS: Add note to generated TAB files
T5SS sectors have canonical data in res/t5ss/data and this is used to
generate the res/Sectors/M1105/*.tab files, pulling in anomalies etc.
People submitting changes should apply them to the source files not
the generated files, but this isn't obvious, leading to extra work for
the maintainer.

Add a text header to the generated files to make this clear, and also
document that blank/comment lines are supported in TSV files.