M1120: Add ZARU by Greg Videll c/o email
Per email:
-----
Back in the day I was with HIWG and later I freelanced for GDW. Dave
Nilsen and I had many discussions about the sector and the fact that
whatever random generator tool that was used for UWPs had a tendency
to "stutter". At Dave's behest I reworked the sector stats to
eliminate the highly repetitive UWPs while at the same time doing as
little damage to Zarushagar as I could. The attachment was the result
of that work and was accepted by Dave; he considered it official even
though it was never published.
-----

Borders don't quite line up, etc.