Reverted nine D worlds in Lishun 1105 and aligned trade remarks with Jim Kundert (#424)
Jim Kundert and I dug into Lishun data to find out exactly why about 1% of the UWPs changed from "D" atmospheres (in Sunbane data) to "A" and a couple "B" atmospheres in the Travellermap 1105 dataset. We traced the work back to my work fixing Sunbane to jibe with AOTI and Traveller5.

The problem is that the data was generated for MegaTraveller originally, which classified "D" atmospheres as Fluid water. This rule was dropped for T5, and as far as we know didn't exist in other rulesets.

So when I fixed Lishun data, instead of just removing the Fl code, I downgraded the atmospheres of ~9 worlds. Fl was a MegaTraveller side-effect of the atmosphere code... not the other 'way 'round.

I have reverted those changes and dropped the Fl codes for those worlds instead. This is a closer fit to the original data and does no harm to AOTI.