feat: adding planet and system generation and extensions to display
This change incorporates:
* Planet and system generation through the use of generate.js
* The generation also supports some additional remark codes:
  * TL:<type>:delta - modify the TL for a subtype of tech
  * LL:<type>:delta - modify the law-level for a sub-aspect
  * !J - Jump to this world is prevented by a close body (usually the star)
  * !M - Gravatic manouver drives do not function this far from the star
  * Inf - Inferno world
  * Inner - Inner world
  * Big - big world
  * Storm - storm world
  * Rad - Radiation world
  * Hsp - Hospitable world
  * Wlt - Wordlet
  * Ice - Ice world

The generation code should follow T5 definitions although implementation may vary slightly.

* Stellar orbit generation is guessed from given
data.  Since all candidates [near, near-companion, etc.]  are equally
likely to contain a star we pick the ones which will be occupied at
random.

* Primary location is guessed based around world remarks.  That is if
the world is marked Tu/Tr or Fr then we assume the world in not in the
HZ orbit.

* If a star is in the expected primary world orbital then we first try
to move the primary to that star.  If that does not work we move it
out one orbital slot.

* We use an X starport rather than a Y spaceport - we assume X means
  that there is nothing there

* The variance in TL in generated worlds does not make sense, instead
we intend to change this to be a modified version of the main TL in
most cases

* Planets outside of the m-drive range will be considered unlikely to
  be populated

* Non-primary Planets with 0-pop will be update to an X spaceport

* We may choose to place starports on promising planets instead of
  spaceports

We may generate additional subworld-entries when a world is a
balkanized govt to represent the different governments